+++ precompiles/staking/tx.go
@@ -54,7 +54,7 @@ func (p Precompile) CreateValidator(
 	method *abi.Method,
 	args []interface{},
 ) ([]byte, error) {
-	msg, delegatorHexAddr, err := NewMsgCreateValidator(args, p.stakingKeeper.BondDenom(ctx))
+	msg, validatorHexAddr, err := NewMsgCreateValidator(args, p.stakingKeeper.BondDenom(ctx))
 	if err != nil {
 		return nil, err
 	}
@@ -64,15 +64,14 @@ func (p Precompile) CreateValidator(
 		"method", method.Name,
 		"commission", msg.Commission.String(),
 		"min_self_delegation", msg.MinSelfDelegation.String(),
-		"delegator_address", delegatorHexAddr.String(),
-		"validator_address", msg.ValidatorAddress,
+		"validator_address", validatorHexAddr.String(),
 		"pubkey", msg.Pubkey.String(),
 		"value", msg.Value.Amount.String(),
 	)
 
 	// we only allow the tx signer "origin" to create their own validator.
-	if origin != delegatorHexAddr {
-		return nil, fmt.Errorf(ErrDifferentOriginFromDelegator, origin.String(), delegatorHexAddr.String())
+	if origin != validatorHexAddr {
+		return nil, fmt.Errorf(ErrDifferentOriginFromDelegator, origin.String(), validatorHexAddr.String())
 	}
 
 	// Execute the transaction using the message server
@@ -82,7 +81,7 @@ func (p Precompile) CreateValidator(
 	}
 
 	// Emit the event for the delegate transaction
-	if err = p.EmitCreateValidatorEvent(ctx, stateDB, msg, delegatorHexAddr); err != nil {
+	if err = p.EmitCreateValidatorEvent(ctx, stateDB, msg, validatorHexAddr); err != nil {
 		return nil, err
 	}