+++ CHANGELOG.md
@@ -131,7 +131,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (p256-precompile) [#2110](https://github.com/evmos/evmos/pull/2110) Adjust `p256` precompile address.
 - (upgrade) [#2117](https://github.com/evmos/evmos/pull/2117) Enable Stride and Osmosis outposts in v16 upgrade handler.
 - (osmosis-outpost) [#2109](https://github.com/evmos/evmos/pull/2109) Add Osmosis outpost to available EVM extensions.
-- (staking-precompile) [#2122](https://github.com/evmos/evmos/pull/2122) Replace bech32 address with evm hex address for `Createvalidator` function.
+- (staking-precompile) [#2122](https://github.com/evmos/evmos/pull/2122) Replace bech32 address with EVM hex address for `Createvalidator` function.
 
 ### Bug Fixes